YOLO_NAS_Course_streamlit/requirements.txt
Cython>=0.29.24
git+https://github.com/philferriere/cocoapi.git#subdirectory=PythonAPI
opencv-python==4.5.5.62
super-gradients
streamlit==1.12.2
cmake==3.22.3
numpy>=1.21.2
torch>=1.9.0
tqdm>=4.57.0
boto3>=1.17.15
jsonschema>=3.2.0
Deprecated>=1.2.11
scipy>=1.6.1
matplotlib>=3.3.4
psutil>=5.8.0
tensorboard>=2.4.1
setuptools>=65.5.1
coverage~=5.3.1
torchvision>=0.10.0
sphinx~=4.0.2
sphinx-rtd-theme
torchmetrics==0.8
hydra-core>=1.2.0
omegaconf
onnxruntime==1.15.0
onnx==1.15.0
pillow>=10.2.0
pip-tools>=6.12.1
einops==0.3.2
protobuf==3.20.3
treelib==1.6.1
termcolor==1.1.0
packaging>=20.4
wheel>=0.38.0
pygments>=2.7.4
stringcase>=1.2.0
